+++ src/routes/Account/components/Prefs/Prefs.js
@@ -4,17 +4,15 @@ import PlayerPrefs from './PlayerPrefs'
 import YouTubePrefs from './YouTubePrefs'
 import styles from './Prefs.css'
 
-export default class Prefs extends React.Component {
-  render () {
-    return (
-      <div className={styles.container}>
-        <h1 className={styles.title}>Preferences</h1>
-        <div className={styles.content}>
-          <PathPrefs />
-          <YouTubePrefs />
-          <PlayerPrefs />
-        </div>
-      </div>
-    )
-  }
-}
+const Prefs = props => (
+  <div className={styles.container}>
+    <h1 className={styles.title}>Preferences</h1>
+    <div className={styles.content}>
+      <PathPrefs />
+      <YouTubePrefs />
+      <PlayerPrefs />
+    </div>
+  </div>
+)
+
+export default Prefs