+++ src/routes/Account/selectors/getRoomList.js
@@ -2,12 +2,12 @@ import { createSelector } from 'reselect'
 
 const getResult = (state) => state.rooms.result
 const getEntities = (state) => state.rooms.entities
-const getShowAll = (state) => state.rooms.isShowingAll
+const getFilterStatus = (state) => state.rooms.filterStatus
 
 const getRoomList = createSelector(
-  [getResult, getEntities, getShowAll],
-  (result, entities, isShowingAll) => ({
-    result: result.filter(roomId => entities[roomId].status === 'open' || isShowingAll),
+  [getResult, getEntities, getFilterStatus],
+  (result, entities, status) => ({
+    result: result.filter(roomId => status === false || entities[roomId].status === status),
     entities,
   }))