+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ### New
 - **(app)** User management interface!
 - **(app)** CD+Graphics backgrounds are now content-aware for better visuals and readability
+- **(app)** Media folders can now be prioritized via drag-and-drop
 - **(app)** Summary is now shown following a media scan
+- **(app)** Added ~370 visualizer presets (now 472 total)
 - **(server)** Support for custom URL base path (see `--urlPath` CLI option)
+- **(server)** Reduced load and improved client connection speed via library caching
 
 ### Changed